+++ src/App.css
@@ -1,50 +1,3 @@
-.GameGrid {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  max-width: 500px;
-}
-
-.Status {
-  font-size: 24px;
-  font-weight: bold;
-  margin: 20px auto;
-}
-
-
-.Square {
-  flex: 1 0 30%; 
-  display: inline-block;
-  background-color: #fff;
-  font-size: 60px;
-  font-weight: bold;
-  line-height: 120px;
-  height: 120px;
-  margin: 3px;
-  padding: 0;
-  text-align: center;
-  width: 120px;
-  cursor: pointer;
-  box-shadow: 2px 2px 2px #ccc;
-  transition: background-color 0.3s ease;
-}
-
-.Square:hover {
-  background-color: #F0F0F0;
-}
-
-.Square:focus {
-  outline: none;
-}
-
-.Square:disabled {
-  pointer-events: none;
-}
-
-.kbd-navigation .Square:focus {
-  background: grey;
-}
-
 .App {
   display: flex;
   margin: 0 auto;
@@ -86,49 +39,7 @@
   color: red;
 }
 
-.Descending {
-
-    -ms-box-orient: horizontal;
-  display: -webkit-box;
-  display: -moz-box;
-  display: -ms-flexbox;
-  display: -moz-flex;
-  display: -webkit-flex;
-  display: flex;
-
-  -webkit-flex-direction: column; 
-  flex-direction: column; 
-  float: center;
-}
-
-.Ascending {
-
-  -ms-box-orient: horizontal;
-  display: -webkit-box;
-  display: -moz-box;
-  display: -ms-flexbox;
-  display: -moz-flex;
-  display: -webkit-flex;
-  display: flex;
-
-  -webkit-flex-direction: column-reverse; 
-  flex-direction: column-reverse; 
-  float: center;
-}
-
 #LatestMove {
   font-weight: bold;
   color: green;
-}
-
-.MovesOrder {
-  box-shadow: 2px 2px 2px #ccc;
-  background-color: green;
-  color: white;
-  text-align: center;
-  font-size: 16px;
-  margin-bottom: 5px;
-  height: 2.2em;
-  transition: background-color 0.3s ease;
-  text-decoration: none;  
 }
