+++ src/components/Square.css
@@ -1,0 +1,32 @@
+.Square {
+  flex: 1 0 30%; 
+  display: inline-block;
+  background-color: #fff;
+  font-size: 60px;
+  font-weight: bold;
+  line-height: 120px;
+  height: 120px;
+  margin: 3px;
+  padding: 0;
+  text-align: center;
+  width: 120px;
+  cursor: pointer;
+  box-shadow: 2px 2px 2px #ccc;
+  transition: background-color 0.3s ease;
+}
+
+.Square:hover {
+  background-color: #F0F0F0;
+}
+
+.Square:focus {
+  outline: none;
+}
+
+.Square:disabled {
+  pointer-events: none;
+}
+
+.kbd-navigation .Square:focus {
+  background: grey;
+}
