+++ src/components/Status.css
@@ -1,0 +1,5 @@
+.Status {
+  font-size: 24px;
+  font-weight: bold;
+  margin: 20px auto;
+}
