+++ src/components/calculateWinner.js
@@ -1,0 +1,25 @@
+const calculateWinner = (squares) => {
+  const lines = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [2, 4, 6]
+  ];
+  for (let i = 0; i < lines.length; i++) {
+    const [a, b, c] = lines[i];
+    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+      const elems = document.getElementsByClassName("Square");
+      for(let i = 0; i < elems.length; i++) {
+        elems[i].disabled = true;
+      }
+      return lines[i];
+    }
+  }
+  return null;
+}
+
+export default calculateWinner;
