+++ src/containers/Board.css
@@ -1,0 +1,6 @@
+.GameGrid {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  max-width: 500px;
+}
